chore: Refactor controller invoker to use event classes for before and after execution
