+++ src/Http/Event/AfterControllerExecution.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Takemo101\Chubby\Http\Bridge;
+namespace Takemo101\Chubby\Http\Event;
 
 use Psr\Http\Message\ResponseInterface;
 use Takemo101\Chubby\Event\StoppableEvent;
